pyproject.toml
[project]
name = "nnunet-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.132.0",
    "locust>=2.43.3",
    "nnunetv2>=2.6.4",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pydantic>=2.12.5",
    "pydicom>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "uvicorn>=0.41.0",
    "websocket-client>=1.9.0",
    "websockets>=16.0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
